#1274 move TriggerKeyEntity to its own file
