+++ app/src/main/java/io/github/sds100/keymapper/data/migration/Migration1To2.kt
@@ -13,7 +13,7 @@ import com.google.gson.JsonArray
 import com.google.gson.JsonObject
 import com.google.gson.JsonParser
 import io.github.sds100.keymapper.data.entities.ActionEntity
-import io.github.sds100.keymapper.data.entities.TriggerEntity
+import io.github.sds100.keymapper.data.entities.TriggerKeyEntity
 import splitties.bitflags.hasFlag
 import timber.log.Timber
 
@@ -107,7 +107,7 @@ object Migration1To2 {
 
                             createTriggerKey2(
                                 it.asInt,
-                                TriggerEntity.KeyEntity.DEVICE_ID_ANY_DEVICE,
+                                TriggerKeyEntity.DEVICE_ID_ANY_DEVICE,
                                 clickType,
                             )
                         }
