+++ app/src/test/java/io/github/sds100/keymapper/data/repositories/KeyMapRepositoryTest.kt
@@ -6,6 +6,7 @@ import io.github.sds100.keymapper.data.entities.ActionEntity
 import io.github.sds100.keymapper.data.entities.Extra
 import io.github.sds100.keymapper.data.entities.KeyMapEntity
 import io.github.sds100.keymapper.data.entities.TriggerEntity
+import io.github.sds100.keymapper.data.entities.TriggerKeyEntity
 import io.github.sds100.keymapper.system.devices.FakeDevicesAdapter
 import io.github.sds100.keymapper.system.devices.InputDeviceInfo
 import io.github.sds100.keymapper.util.State
@@ -254,7 +255,7 @@ class KeyMapRepositoryTest {
     fun `key map with device name for trigger key, if device for trigger key is connected, do not update trigger key device name`() =
         runTest(testDispatcher) {
             // GIVEN
-            val triggerKey = TriggerEntity.KeyEntity(
+            val triggerKey = TriggerKeyEntity(
                 keyCode = 1,
                 deviceId = FAKE_KEYBOARD.descriptor,
                 deviceName = FAKE_KEYBOARD.name,
@@ -277,7 +278,7 @@ class KeyMapRepositoryTest {
     fun `key map with device name for trigger key, if device for trigger key is disconnected, do not update trigger key device name`() =
         runTest(testDispatcher) {
             // GIVEN
-            val triggerKey = TriggerEntity.KeyEntity(
+            val triggerKey = TriggerKeyEntity(
                 keyCode = 1,
                 deviceId = FAKE_KEYBOARD.descriptor,
                 deviceName = FAKE_KEYBOARD.name,
@@ -298,7 +299,7 @@ class KeyMapRepositoryTest {
     fun `key map with no device name for trigger key, if device for trigger key is connected, update trigger key device name`() =
         runTest(testDispatcher) {
             // GIVEN
-            val triggerKey = TriggerEntity.KeyEntity(
+            val triggerKey = TriggerKeyEntity(
                 keyCode = 1,
                 deviceId = FAKE_KEYBOARD.descriptor,
                 deviceName = "",
